pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "maze-esolang"
version = "1.0.0"
description = "A Python interpreter and visual debugger for the 2D esoteric programming language, Maze."
readme = "README.md"
authors = [
    {name = "Yosi Schlakman"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "pygame-ce>=2.4.0"
]

[project.scripts]
maze = "maze.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]